requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fastapi==0.135.2
fonttools==4.62.1
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.18.1
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pillow==12.1.1
plotly==6.6.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
pywebview==5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
# uvloop 不支持 Windows；CI 的 Windows runner 会因此安装失败
uvloop==0.22.1; platform_system != "Windows"
watchfiles==1.1.1
websockets==16.0
